Fix cargo clippy

+++ pallets/parachain-staking/src/lib.rs
@@ -2637,10 +2637,10 @@ pub mod pallet {
 			if let Some(state) = CandidatePool::<T>::get(author) {
 				let pot = Self::account_id();
 				let ed = <T::Currency as frame_support::traits::fungible::Inspect<T::AccountId>>::minimum_balance();
-				let issue_number = if ed == T::CurrencyBalance::from(0 as u32) {
+				let issue_number = if ed == T::CurrencyBalance::from(0_u32) {
 					T::Currency::reducible_balance(&pot, Preservation::Preserve, Fortitude::Polite)
 						// Avoid the pot complaint no balance there
-						.checked_sub(&T::CurrencyBalance::from(10 as u32))
+						.checked_sub(&T::CurrencyBalance::from(10_u32))
 						.unwrap_or_else(Zero::zero)
 				} else {
 					T::Currency::reducible_balance(&pot, Preservation::Preserve, Fortitude::Polite)
